Replaces growpartfs with the systemd equivalent functionalities.
This configures systemd-repart to resize the root partition, and
systemd-growfs to resize the root filesystem.

Both services run early in the boot process, and expect the /dev
filesystem to be properly populated. This was only happening later in
the bootup process in the previous configuration, as udev was not
started in the minimal initrd that was used, and as a result they would
both fail.

To fix this, systemd is now enabled in initrd, so that udev is started
there and we get a more complete system earlier. This makes initrd
somewhat bigger, but it probably makes the setup more robust in
general.